move user & dashboard to model
