+++ pkg/model/user.go
@@ -1,4 +1,4 @@
-package auth
+package model
 
 import "time"
