+++ pkg/store/dashboard_store.go
@@ -1,13 +1,14 @@
 package store
 
 import (
+	"github.com/kuoss/venti/pkg/model"
 	"log"
 	"os"
 	"path/filepath"
 )
 
 type DashboardStore struct {
-	dashboards []Dashboard
+	dashboards []model.Dashboard
 }
 
 // NewDashboardStore pattern parameter is root filepath pattern for dashboard yaml files. ex) etc/dashboards/**/*.yaml
@@ -18,14 +19,14 @@ func NewDashboardStore(pattern string) (*DashboardStore, error) {
 	if err != nil {
 		return nil, err
 	}
-	dashboards := make([]Dashboard, 0)
+	dashboards := make([]model.Dashboard, 0)
 	for _, filename := range files {
 		log.Printf("dashboard file: %s\n", filename)
 		f, err := os.Open(filename)
 		if err != nil {
 			log.Printf("error open dashboard file: %s\n", err.Error())
 		}
-		var d *Dashboard
+		var d *model.Dashboard
 		err = loadYaml(f, d)
 		if err != nil {
 			return nil, err
@@ -35,6 +36,6 @@ func NewDashboardStore(pattern string) (*DashboardStore, error) {
 	return &DashboardStore{dashboards: dashboards}, nil
 }
 
-func (dbs *DashboardStore) Dashboards() []Dashboard {
+func (dbs *DashboardStore) Dashboards() []model.Dashboard {
 	return dbs.dashboards
 }
